Refactor generation services transport layers
