+++ app/frontend/src/composables/generation/useGenerationOrchestrator.ts
@@ -1,7 +1,5 @@
-import type {
-  GenerationQueueClient,
-  GenerationWebSocketManager,
-} from '@/services';
+import type { GenerationQueueClient } from '@/services/generation/queueClient';
+import type { GenerationWebSocketManager } from '@/services/generation/websocketManager';
 
 import type { GenerationNotificationAdapter } from './useGenerationTransport';
 import {
