+++ app/frontend/src/composables/generation/useGenerationTransport.ts
@@ -1,8 +1,6 @@
-import {
-  extractGenerationErrorMessage,
-  type GenerationQueueClient,
-  type GenerationWebSocketManager,
-} from '@/services';
+import { extractGenerationErrorMessage } from '@/services/generation/updates';
+import { type GenerationQueueClient } from '@/services/generation/queueClient';
+import { type GenerationWebSocketManager } from '@/services/generation/websocketManager';
 import type {
   GenerationCompleteMessage,
   GenerationErrorMessage,
