+++ app/frontend/src/services/generation/index.ts
@@ -1,3 +1,5 @@
 export * from './updates';
+export * from './queueClient';
+export * from './websocketManager';
 export * from './generationService';
 export * from './validation';
